requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1
schedule==1.2.0
pytz==2023.3
selenium==4.15.2
webdriver-manager==4.0.1
mcp>=1.9.2 